enable buffering for the IPC channel
I always intended it to be this way, but forgot to actually enable it.
During times where the cc-slurm-adapter timer executing, there may be
some delay when the ipcSocketChan is emptied.
In order to keep delays for the slurmctld PrEp hook as low as possible,
any notifications should be buffered.