mark jobs as stale, before actually reporting them as stale
This fixes a 'race' condition, where jobs were reported as stale, even
though they simply finished since the last poll, but weren't marked as
stopped yet.

In our case, mark them as stale first, and if they are still stale
during the next poll, then there is a problem. This case is handled as
previously.