GRES: Only match the first valid host regex
Otherwise, if a host appears in multiple regexes, the PCI address from
both hosts patterns will be added to the job.

This is undesirable in any case. With the this fix, we can also have
"overrides", where a more specific regex can be used before a more
generic one to make "exceptions".